Query available rooms completed, hypermedia not working

+++ src/main/java/com/esi/bookings/application/dto/PropertyDTO.java
@@ -17,7 +17,7 @@ public class PropertyDTO extends ResourceSupport{
 
     String city;
 
-    BigDecimal price;
+    int price;
 
     String address;
